src/vk/requirements.txt
python-decouple~=3.8
flask[async]~=3.0.2
requests==2.31.0
